Safecopy support in ethernet drivers.

+++ drivers/dpeth/devio.c
@@ -29,7 +29,7 @@ static void warning(const char *type, int err)
 */
 PUBLIC unsigned int inb(unsigned short port)
 {
-  unsigned int value;
+  unsigned long value;
   int rc;
 
   if ((rc = sys_inb(port, &value)) != OK) warning("inb", rc);
@@ -42,7 +42,7 @@ PUBLIC unsigned int inb(unsigned short port)
 */
 PUBLIC unsigned int inw(unsigned short port)
 {
-  unsigned int value;
+  unsigned long value;
   int rc;
 
   if ((rc = sys_inw(port, &value)) != OK) warning("inw", rc);
